requirements.txt
pillow
qrcode
pyzipper
numpy
base58